0.0.6

+++ Gradio.Net/Components/Component.cs
@@ -24,7 +24,7 @@ namespace Gradio.Net
 
         internal object Value { get; set; }
         internal string Label { get; set; }
-        internal float? Every { get; set; }
+        internal decimal? Every { get; set; }
         internal bool? ShowLabel { get; set; }
